+++ Marlin/src/lcd/ultralcd.cpp
@@ -150,7 +150,7 @@ constexpr uint8_t epps = ENCODER_PULSES_PER_STEP;
     volatile uint8_t MarlinUI::slow_buttons;
   #endif
   #if HAS_TOUCH_XPT2046
-    #include "touch/xpt2046.h"
+    #include "touch/touch_buttons.h"
     bool MarlinUI::on_edit_screen = false;
   #endif
 #endif
